tinisoft/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.0
python-dotenv==1.0.0
requests==2.31.0
dnspython==2.4.2
docker==7.0.0
PyJWT==2.8.0
cryptography==42.0.2
gunicorn==21.2.0
whitenoise==6.6.0
django-environ==0.11.2

